+++ tests/integration/ssr/fixtures/base/modern.config.ts
@@ -7,7 +7,9 @@ export default defineConfig({
     router: true,
   },
   server: {
-    ssr: true,
+    ssr: {
+      disablePrerender: true,
+    },
   },
   tools: {
     webpack(config) {