+++ packages/devtools/plugin/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @modern-js/plugin-devtools
 
+## 2.34.0
+
+### Patch Changes
+
+- Updated dependencies [a77b82a]
+- Updated dependencies [dcdeea5]
+- Updated dependencies [eca24a2]
+- Updated dependencies [c8b448b]
+  - @modern-js/utils@2.34.0
+  - @modern-js/runtime@2.34.0
+  - @modern-js/devtools-client@2.34.0
+  - @modern-js/devtools-mount@2.34.0
+
 ## 2.33.1
 
 ### Patch Changes