+++ packages/runtime/plugin-router-v5/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/plugin-router-v5
 
+## 2.34.0
+
+### Patch Changes
+
+- Updated dependencies [a77b82a]
+- Updated dependencies [c8b448b]
+  - @modern-js/utils@2.34.0
+  - @modern-js/plugin@2.34.0
+  - @modern-js/types@2.34.0
+
 ## 2.33.1
 
 ### Patch Changes